refactor: type을 하나로 관리
