+++ server/src/types/index.ts
@@ -1,1 +1,3 @@
 export * from './bookmark';
+export * from './comment';
+export * from './user';
